Add CI job for the UI component
test coverage is rather weak and we are not planning on using this
component but it's good to have at least minimal coverage which will
help spot incompatibilities with dependencies for example.

This is a minimum effort to have the UI component covered in CI in case
we ever need it!